requirements.txt
Flask==2.3.3
pandas==2.0.3
praw==7.7.1
python-dotenv==1.0.0
openpyxl==3.1.2
XlsxWriter==3.1.9